Move camp::Dictionary to camp::detail
